requirements.txt
Django
python-dotenv
django-environ
psycopg2-binary
django-admin-relation-links
whitenoise
django-pwa
